feat: game engine implementation part 1

+++ Project/MinistryOfTruth/Domain/Models/GameState.cs
@@ -6,8 +6,8 @@ public record GameState(
     int Day,
     int Score,
     int TextsRemaining,
-    float TimeLeftRatio, // 0-1
-    float DangerLevelRatio, // 0-1
+    double TimeLeftRatio, // 0-1
+    double DangerLevelRatio, // 0-1
     bool IsCorrectDecision = true,
     string StatusMessage = ""
 );
